+++ src/components/main/themeBanner/ThemeSlideBanner.css
@@ -1,0 +1,45 @@
+/* 슬라이드 CSS */
+
+.swiper {
+  width: 100%;
+  height: 100%;
+}
+
+.swiper-slide {
+  text-align: center;
+  font-size: 18px;
+  background: #fff;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.swiper-button-next {
+  width: 30px;
+  opacity: 0.5;
+  background-size: 50% auto;
+  background-position: center;
+}
+
+.swiper-button-next:hover {
+  opacity: 1;
+}
+
+.swiper-button-prev {
+  width: 50px;
+  background-size: 50% auto;
+  background-position: center;
+  opacity: 0.5;
+}
+.swiper-button-prev:hover {
+  opacity: 1;
+}
+.swiper-button-next::after,
+.swiper-button-prev::after {
+  /* display: none; */
+  color: #797979;
+}
+
+.swiper-pagination-bullet {
+  background-color: #797979;
+}
